changes for user checking in both comment and alert

+++ backend/comments/views.py
@@ -50,13 +50,22 @@ def createComment(request):
 def get_comments_for_paper(request, paper_id):
     if request.method == 'GET':
        try:
+        print(request.body)
+        username=request.data.get('user')
         pid=paper_id
-        print(pid)
+       
+        #print(pid)
         paper_info = get_object_or_404(PaperInfo, paperId=pid)
-        print(paper_info)
-        comments = CommentsCache.objects.filter(paper_id=pid).values('comment_id', 'user', 'text')  # Add other fields as needed
-
+       # print(paper_info)
+        comments = CommentsCache.objects.filter(paper_id=pid,user=username).values('_id', 'user', 'text','keyword')  # Add other fields as needed
+       # comments.queryset
+        
+        for c in range(len(comments)):
+            comments[c]['_id']=str(comments[c]['_id'])
+        #print(comments)
         return Response(comments, status=status.HTTP_200_OK)
+       except CommentsCache.DoesNotExist:
+           return Response({'error': f'User with username {username} does not exist'}, status=status.HTTP_404_NOT_FOUND)
        except PaperInfo.DoesNotExist:
         return Response({'error': f'PaperInfo with paper_id {paper_id} does not exist'}, status=status.HTTP_404_NOT_FOUND)
 
@@ -68,7 +77,8 @@ def get_comments_for_paper(request, paper_id):
 def update_comment(request,comment_id):
     if request.method == 'PUT':
         try:
-            comment = CommentsCache.objects.get(_id=ObjectId(comment_id))
+            username=request.data.user('user')
+            comment = CommentsCache.objects.filter(_id=ObjectId(comment_id),user=username).values('_id', 'user', 'text','keyword')
         except CommentsCache.DoesNotExist:
             return Response({'error': f'Comment with comment_id {comment_id} does not exist'}, status=status.HTTP_404_NOT_FOUND)
 
@@ -86,14 +96,19 @@ def update_comment(request,comment_id):
 @api_view(['DELETE'])
 def delete_comment(request, comment_id):
     if request.method == 'DELETE':
+        
         try:
-            
-            comment = CommentsCache.objects.get(_id=ObjectId(comment_id))
-            print(comment)
+            print(request.body)
+            username=request.data.get('user')
+            print(username)
+            comment = CommentsCache.objects.filter(_id=ObjectId(comment_id),user=username)
+            if comment:
+                print(comment)
+                comment.delete()
         except CommentsCache.DoesNotExist:
             return Response({'error': f'Comment with comment_id {comment_id} does not exist'}, status=status.HTTP_404_NOT_FOUND)
 
-        comment.delete()
+        
         return Response({'success': 'Comment deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
 
     return Response({'error': 'Only DELETE method is supported'}, status=status.HTTP_400_BAD_REQUEST)
